requirements.txt
# ============================================
# CHRONOS FORECASTER - Requirements
# ============================================

# Core ML
chronos-forecasting>=1.2.0
torch>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# API Server
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Utilities
openpyxl>=3.1.0  # Per file Excel
